+++ src/WorkOS/Admin/LoginProfiles/AdminPage.php
@@ -95,6 +95,8 @@ class AdminPage {
 			true
 		);
 
+		wp_set_script_translations( self::SCRIPT_HANDLE, 'integration-workos' );
+
 		wp_enqueue_style(
 			self::STYLE_HANDLE,
 			$assets_url . 'admin-profiles.css',